Separate circuit.py into three files
