+++ scqubits/__init__.py
@@ -31,7 +31,7 @@ from scqubits.core.transmon import Transmon, TunableTransmon
 from scqubits.core.zeropi import ZeroPi
 from scqubits.core.zeropi_full import FullZeroPi
 # arbitrary circuit stuff
-from scqubits.core.circuit import Circuit, CircuitElement, CircuitNode, Capacitance, Inductance, JosephsonJunction, Variable
+from scqubits.core.circuit import Circuit, CircuitNode, Variable, ExternalFlux, ExternalCharge, Capacitance, JosephsonJunction, Inductance, LagrangianCurrentSource
 from scqubits.core.circuit_flux_qubit import CircuitFluxQubit
 # file IO
 from scqubits.io_utils.fileio import read, write
